rag/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
jinja2==3.1.4
python-multipart==0.0.12

langchain==0.3.0
langchain-openai==0.2.0
langchain-milvus==0.1.6
langgraph==0.2.0

pymilvus==2.4.0

python-dotenv==1.0.1
tavily-python==0.5.0

# 可选：用于绘制图形
pygraphviz==1.13
